Added json to test
